+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 2024-05-03
 * Add support for analyzing dependencies managed by Swift Package Manager.
 * Add ios spm example.
+* Add `-i` command line option.
 
 ## 2024-05-01
 * Improve the accuracy of CocoaPods project analysis.